docs(barchart): fix examples structure
