--- _examples/barchart/barchart_001.go
@@ -1,28 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"github.com/pterm/pterm"
-)
-
-func main() {
-	positiveBars := pterm.Bars{
-		pterm.Bar{
-			Label: "Bar 1",
-			Value: 5,
-		},
-		pterm.Bar{
-			Label: "Bar 2",
-			Value: 3,
-		},
-		pterm.Bar{
-			Label: "Longer Label",
-			Value: 7,
-		},
-	}
-
-	fmt.Println("Chart example with positive only values (bars use 100% of chart area)")
-	_ = pterm.DefaultBarChart.WithBars(positiveBars).Render()
-	_ = pterm.DefaultBarChart.WithHorizontal().WithBars(positiveBars).Render()
-	fmt.Println("=====================================================================")
-}
