--- _examples/barchart/barchart_003.go
@@ -1,36 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"github.com/pterm/pterm"
-)
-
-func main() {
-	mixedBars := pterm.Bars{
-		pterm.Bar{
-			Label: "Bar 1",
-			Value: 2,
-		},
-		pterm.Bar{
-			Label: "Bar 2",
-			Value: -3,
-		},
-		pterm.Bar{
-			Label: "Bar 3",
-			Value: -2,
-		},
-		pterm.Bar{
-			Label: "Bar 4",
-			Value: 5,
-		},
-		pterm.Bar{
-			Label: "Longer Label",
-			Value: 7,
-		},
-	}
-
-	fmt.Println("Chart example with mixed values (note screen space usage in case when ABSOLUTE values of negative and positive parts are differ too much)")
-	_ = pterm.DefaultBarChart.WithBars(mixedBars).Render()
-	_ = pterm.DefaultBarChart.WithHorizontal().WithBars(mixedBars).Render()
-	fmt.Println("=====================================================================")
-}
